requirements.txt
discord.py
python-dotenv
sqlalchemy
psycopg2-binary